+++ example/README.md
@@ -34,8 +34,14 @@ The headings below describe the functionality you can find in each subview acces
 
 ### Query Notifications
 
-- Demonstrates usage of `b`. Modify the `NotificationHistoryView.pageView` function to experiment with different query parameters
+- Demonstrates usage of `ParticipantSession.queryNotifications`. Modify the `NotificationHistoryView.pageView` function to experiment with different query parameters. `NotificationHistoryView.Model` demonstrates various ways to access the different types of notification content available
 
 ### MyDataHelps Embeddable Surveys
 
 To see examples of using `EmbeddableSurveyViewController` to present MyDataHelps embeddable surveys, tap on incomplete surveys in the Survey Tasks screen. To use this feature, embeddable survey functionality must be enabled in RKStudio for the project and the survey. See EmbeddableSurveyViewController's documentation for more info on usage and links to relevant RKStudio documentation.
+
+### External Accounts
+
+- The top level External Accounts screen demonstrates usage of `ParticipantSession.listExternalAccounts` to view, update, and delete connected external account providers
+- Tap the `+` button on that screen to view available external account providers via `ParticipantSession.queryExternalAccountProviders`. Modify `ProvidersListViewModel` to experiment with different query parameters
+- Selecting an external account provider demonstrates the provider connection authorization flow, using `ParticipantSession.connectExternalAccount` to initiate the connection, and `SFSafariViewController` to present the UI. Note that MyDataHelpsKit only supplies a URL to present, your app (as demonstrated by the example app) is responsible for presenting a Safari view with that URL to the participant, and dismissing the Safari view by intercepting a special link. See `ParticipantSession.connectExternalAccount` documentation for details
